+++ Practice/MLoginova/Loginova_10.1.py
@@ -2,28 +2,15 @@
 #  start (по умолчанию 3) до заданного числа end. Далее необходимо:
 #  Запустить ее три раза последовательно в диапазоне от 3 до 10000, от 10001 до 20000, от 20001 до 30000.
 import time
-
-
-def find_primes(start=3, end=None):
-    pr_chisla = []
-    while start <= end:
-        k = 2
-        while k <= start**0.5:
-            if start % k == 0:
-                break
-            k += 1
-        else:
-            pr_chisla.append(start)
-        start += 1
-    #  print(pr_chisla)
+import Loginova_10
 
 
 if __name__ == "__main__":
     data = [(3, 10000), (10001, 20000), (20001, 30000)]
     start = time.perf_counter()
     for i in range(len(data)):
-        find_primes(*data[i])
-        print(f'{i+1}.Время вычислений в секундах для диапазона {data[i]}: {time.perf_counter() - start}')
-#  1.Время вычислений в секундах для диапазона (3, 10000): 0.017632799921557307
-#  2.Время вычислений в секундах для диапазона (10001, 20000): 0.04520189994946122
-#  3.Время вычислений в секундах для диапазона (20001, 30000): 0.0779491999419406
+        Loginova_10.find_primes(*data[i])
+        print(f'{i+1}.Время вычислений в секундах: {time.perf_counter() - start}')
+#  1.Время вычислений в секундах: 0.007450300035998225
+#  2.Время вычислений в секундах: 0.018060300033539534
+#  3.Время вычислений в секундах: 0.03094460000284016
