+++ Practice/MLoginova/Loginova_10.1_threading.py
@@ -1,26 +1,13 @@
 import threading
 import time
-
-def find_primes(start=3, end=None):
-    pr_chisla = []
-    while start <= end:
-        k = 2
-        while k <= start**0.5:
-            if start % k == 0:
-                break
-            k += 1
-        else:
-            pr_chisla.append(start)
-        start += 1
-    #  print(pr_chisla)
-
+import Loginova_10
 
 if __name__ == "__main__":
     data = [(3, 10000), (10001, 20000), (20001, 30000)]
     start = time.perf_counter()
     threads = []
     for i in range(len(data)):
-        thr = threading.Thread(target=find_primes, args=(*data[i],))
+        thr = threading.Thread(target=Loginova_10.find_primes, args=(*data[i],))
         thr.start()
         # Если забыть выполнить start для потоков, то мы не направим ОС команду на создание потока,
         # код вернёт исключение RuntimeError: cannot join thread before it is started
@@ -32,6 +19,6 @@ if __name__ == "__main__":
         #  программа завершит свое выполнение, не дожидаясь завершения наших потоков в итоге замеряемое время
         #  для каждого потока будет неккоректным
         print(f'Время вычислений c использованием потоков в секундах: {time.perf_counter() - start}')
-#  1.Время вычислений c использованием потоков в секундах: 0.05176529998425394
-#  2.Время вычислений c использованием потоков в секундах: 0.08369849994778633
-#  3.Время вычислений c использованием потоков в секундах: 0.10709089995361865
+#  1.Время вычислений c использованием потоков в секундах: 0.03124839998781681
+#  2.Время вычислений c использованием потоков в секундах: 0.03127939999103546
+#  3.Время вычислений c использованием потоков в секундах: 0.031286600045859814
